DSC-203 added header from fe-common
