+++ src/components/layout/styles/Layout.scss
@@ -15,3 +15,8 @@ footer a img {
 .dojo-footer .container {
   max-width: 1220px;
 }
+
+.header{
+  background: white;
+  box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.08), 0px 1px 4px rgba(0, 0, 0, 0.08);
+}
