--- src/components/site-header/index.js
@@ -1,1 +1,0 @@
-export { default as SiteHeader } from './SiteHeader';
