--- src/components/site-header/menu/index.js
@@ -1,3 +1,0 @@
-import { Menu, MenuTrigger, MenuContent } from './Menu';
-
-export { Menu, MenuTrigger, MenuContent };
