--- src/components/site-header/styles/Menu.scss
@@ -1,44 +1,0 @@
-.menu {
-  position: relative;
-}
-.menu-content {
-  position: absolute;
-  top: 100%;
-  z-index: 10;
-  background: $white;
-  min-width: 10rem;
-  &.pin-left {
-    left: 0;
-  }
-  &.pin-right {
-    right: 0;
-  }
-}
-
-.menu-dropdown-enter {
-  opacity: 0;
-  transform-origin: 75% 0;
-  transform: scale3d(0.8, 0.8, 1);
-}
-.menu-dropdown-enter-active {
-  transform-origin: 75% 0;
-  transition: all 250ms cubic-bezier(0.4, 0, 0.2, 1);
-  transform: scale3d(1, 1, 1);
-  opacity: 1;
-}
-.menu-dropdown-enter-done {
-}
-
-.menu-dropdown-exit {
-  transform-origin: 75% 0;
-  transform: scale3d(1, 1, 1);
-  opacity: 1;
-}
-.menu-dropdown-exit-active {
-  transform-origin: 75% 0;
-  transform: scale3d(0.8, 0.8, 1);
-  transition: all 250ms cubic-bezier(0.8, 0, 0.6, 1);
-  opacity: 0;
-}
-.menu-dropdown-exit-done {
-}
