--- src/components/site-header/styles/_Header.scss
@@ -1,103 +1,0 @@
-@import './Menu.scss';
-
-.site-header-mobile,
-.site-header-desktop {
-  background: $white;
-  position: relative;
-  z-index: 1000;
-  .dropdown-item a {
-    text-decoration: none;
-  }
-  a.logo img {
-    max-height: 50px;
-  }
-}
-
-.site-header-mobile {
-  .nav-link {
-    text-decoration: none;
-    cursor: pointer;
-    &.active {
-      color: $component-active-color;
-      background-color: $component-active-bg;
-    }
-  }
-  img {
-    height: 1.5rem;
-  }
-  .icon-button {
-    display: inline-flex;
-    line-height: 3rem;
-    background: transparent;
-    vertical-align: middle;
-    text-align: center;
-    border: none;
-    height: 3rem;
-    width: 3rem;
-    padding: 0.75rem;
-    justify-content: center;
-    align-items: center;
-    &:hover,
-    &:focus {
-      background: rgba(0, 0, 0, 0.1);
-    }
-  }
-}
-
-.site-header-desktop {
-  @extend .border-bottom;
-  background: $white;
-  .nav-link {
-    text-decoration: none;
-  }
-  .logo {
-    display: block;
-    box-sizing: content-box;
-    position: relative;
-    top: -0.05em;
-    height: 1.75rem;
-    padding: 1rem 0;
-    margin-right: 1rem;
-    img {
-      display: block;
-      height: 100%;
-    }
-  }
-  .main-nav {
-    .nav-link {
-      padding: 1.125rem 1rem;
-      text-decoration: none;
-      font-weight: 500;
-      letter-spacing: 0.01em;
-    }
-    .nav-link:hover,
-    .nav-link:focus,
-    .expanded .nav-link {
-      background: $component-active-bg;
-      color: $component-active-color;
-    }
-    .nav-link.active {
-      border-bottom: 4px solid $component-active-bg;
-    }
-    .menu {
-      position: static;
-      .menu-content {
-        border-top: solid 2px $component-active-bg;
-        left: 0;
-        right: 0;
-        box-shadow: 0 1px 2px rgba(0, 0, 0, 0.25);
-        border-bottom-left-radius: 2px;
-        border-bottom-right-radius: 2px;
-        padding: 1rem;
-      }
-    }
-  }
-}
-
-.skip-nav-link:active,
-.skip-nav-link:focus {
-  position: relative;
-  z-index: 1010;
-  padding: $input-btn-padding-y-sm $input-btn-padding-x-sm;
-  font-size: $font-size-sm;
-}
