--- src/components/site-header/tests/SiteHeader.test.jsx
@@ -1,79 +1,0 @@
-import React from 'react';
-import { screen } from '@testing-library/react';
-import '@testing-library/jest-dom/extend-expect';
-import { breakpoints } from '@edx/paragon';
-import userEvent from '@testing-library/user-event';
-
-import { AppContext } from '@edx/frontend-platform/react';
-import SiteHeader from '../SiteHeader';
-
-import { renderWithRouter } from '../../../utils/tests';
-
-const appState = {
-  enterpriseConfig: {
-    name: 'BearsRUs',
-    uuid: 'BearsRUs',
-    slug: 'bears-r-us',
-    branding: {
-      logo: 'the-logo',
-    },
-  },
-  config: {
-    LMS_BASE_URL: process.env.LMS_BASE_URL,
-  },
-  authenticatedUser: {
-    username: 'papa',
-    profileImage: 'papa-bear-image',
-  },
-};
-
-/* eslint-disable react/prop-types */
-const SiteHeaderWithContext = ({ initialAppState = appState }) => (
-  <AppContext.Provider value={initialAppState}>
-    <SiteHeader />
-  </AppContext.Provider>
-);
-/* eslint-enable react/prop-types */
-
-const mockWindowConfig = {
-  type: 'screen',
-  width: breakpoints.large.minWidth + 1,
-  height: 800,
-};
-
-describe('<SiteHeader />', () => {
-  beforeEach(() => {
-    window.matchMedia.setConfig(mockWindowConfig);
-  });
-  test('renders link with logo to dashboard', () => {
-    renderWithRouter(<SiteHeaderWithContext />);
-    expect(screen.getByTestId('header-logo-image-id'));
-  });
-  test('renders regular logout link in absence of IDP', () => {
-    renderWithRouter(<SiteHeaderWithContext initialAppState={appState} />);
-
-    userEvent.click(screen.getByText('papa'));
-    expect(screen.getByText('Sign out')).toBeInTheDocument();
-    const logoutLink = screen.getByText('Sign out');
-    // note: the values of these come from the process.env vars in setupTest.js
-    expect(logoutLink.getAttribute('href')).toBe('http://localhost:18000/logout?next=http://localhost:8734/');
-  });
-  test('renders logout-specific logout link in presence of IDP', () => {
-    const appStateWithIDP = {
-      ...appState,
-      enterpriseConfig: {
-        ...appState.enterpriseConfig,
-        identityProvider: 'a-provider',
-      },
-    };
-    renderWithRouter(<SiteHeaderWithContext initialAppState={appStateWithIDP} />);
-
-    userEvent.click(screen.getByText('papa'));
-    expect(screen.getByText('Sign out')).toBeInTheDocument();
-    const logoutLink = screen.getByText('Sign out');
-    // note: the values of these come from the process.env vars in setupTest.js
-    expect(logoutLink.getAttribute('href')).toBe(
-      'http://localhost:18000/logout?next=http://localhost:8734/%3Flogout=true',
-    );
-  });
-});
