+++ src/index.scss
@@ -17,7 +17,6 @@ $fa-font-path: '~font-awesome/fonts';
 @import './components/dashboard/styles/DashboardDrawer';
 @import './components/dashboard/styles/DashboardPage';
 @import './components/dashboard/styles/Pagination';
-@import './components/site-header/styles/Header';
 @import './components/error-page/styles/ErrorPage';
 @import './components/filter/styles/Filter';
